Make typeString cursor restoration optional
In some cases (the various upcoming terminal-focused scenarios) it's a
bit problematic because it results in terminal flicker, and it's not
helpful because we're just constantly typing anyway (and when we stop
typing, we switch window which fixes the lost-cursor anyway). Making it
optional lets us work around that more flexibly.